libipt, block: classify instructions at trampolines as ptic_other, if possible
If we cannot reach the next instruction due to space limitations in the
block cache, we insert a trampoline cache entry.  We are not able to store
or infer the instruction class for trampoline entries in general.

We do know, however, that we only insert trampolines at non-decision points.
If we end blocks on direct calls and jumps, we know that this is not a
branch instruction (nor a ptwrite).  Set the instruction class to ptic_other
in that case.

Change-Id: Ieaf0faadf5aae42b4a2c43d399d4a02eed43428f
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>